slight refactor to move functions to better places
